+++ sorts/selection.cpp
@@ -27,7 +27,7 @@ int main() {
         const long percent = 1000 * iteration / iterations;
         cout << "\r\bprogress: " << iteration << " / " << iterations << " (" << percent / 10 << '.' << percent % 10 << "%)";
         size_t local_length = length + iteration;
-        long long measured_time = 0;
+        double measured_time = 0;
 
         for (short _ = 0; _ < averaging; _++) {
             utils::generate_array(array_type, min, max, local_length, array);
